move datetime_to_utc into Util
